+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- `medaka features` fails if encoder expects dwells but data does not contain move tables, or warns if available move tables are not being used.  
+- During training, warning is output if model is not using available dwell information.  
+
 ## [v2.1.1]
 ### Added
 - Python 3.12 support